🚚 move event store append feature to event store crate
