+++ marketplace-event-store/src/domain/mod.rs
@@ -1,0 +1,2 @@
+mod event_store;
+pub use event_store::{Error as EventStoreError, EventStore};
