+++ marketplace-event-store/src/infrastructure/database/models/events.rs
@@ -1,20 +1,11 @@
-use crate::database::schema::*;
 use chrono::NaiveDateTime;
-use rocket::serde::{Deserialize, Serialize};
+use diesel::Insertable;
+use marketplace_infrastructure::database::schema::*;
+use serde::{Deserialize, Serialize};
 use serde_json::Value;
 
-#[derive(Debug, Serialize, Deserialize)]
-#[serde(crate = "rocket::serde")]
-pub struct Event {
-	pub aggregate_name: String,
-	pub aggregate_id: String,
-	pub payload: Value,
-}
-
 #[derive(Insertable, Debug, Serialize, Deserialize)]
-#[table_name = "events"]
-#[serde(crate = "rocket::serde")]
-pub struct NewEvent {
+pub struct Event {
 	pub timestamp: NaiveDateTime,
 	pub aggregate_name: String,
 	pub aggregate_id: String,
@@ -24,7 +15,6 @@ pub struct NewEvent {
 }
 
 #[derive(Insertable, Debug, Serialize, Deserialize)]
-#[serde(crate = "rocket::serde")]
 pub struct EventDeduplication {
 	pub deduplication_id: String,
 	pub event_index: i32,
