+++ marketplace-event-store/src/infrastructure/database/models/mod.rs
@@ -1,0 +1,3 @@
+#[allow(clippy::extra_unused_lifetimes)]
+mod events;
+pub use events::*;
