+++ marketplace-event-store/src/infrastructure/mod.rs
@@ -1,0 +1,4 @@
+mod database;
+
+#[cfg(test)]
+mod tests;
