+++ marketplace-event-store/src/infrastructure/tests.rs
@@ -1,19 +1,21 @@
-use crate::{
-	database::{schema::events, Client},
-	diesel::{query_dsl::select_dsl::SelectDsl, RunQueryDsl},
-};
 use chrono::Utc;
-use marketplace_domain::*;
+use diesel::{query_dsl::select_dsl::SelectDsl, RunQueryDsl};
+use marketplace_domain::{
+	Contribution, ContributionEvent, ContributionId, ContributorAccountAddress, EventOrigin,
+	EventStore as DomainEventStore, HexPrefixedString,
+};
+use marketplace_infrastructure::database::{schema::events, Client};
 use marketplace_tests::init_pool;
-use rocket::serde::json::serde_json::Value;
 use rstest::{fixture, rstest};
-use serde_json::json;
+use serde_json::{json, Value};
 use std::str::FromStr;
 
+use crate::{domain::EventStore, Event};
+
 #[fixture]
-fn event_store() -> Box<dyn marketplace_domain::EventStore<Contribution>> {
-	let event_store = Client::new(init_pool());
-	Box::new(event_store) as Box<dyn EventStore<Contribution>>
+#[once]
+fn database() -> Client {
+	Client::new(init_pool())
 }
 
 #[fixture]
@@ -27,14 +29,15 @@ fn contributor_account_address() -> ContributorAccountAddress {
 }
 
 #[fixture]
-fn creation_event(contribution_id: ContributionId) -> StorableEvent<Contribution> {
-	StorableEvent {
+fn creation_event(contribution_id: ContributionId) -> Event {
+	Event {
 		event: ContributionEvent::Created {
 			id: contribution_id,
 			project_id: Default::default(),
 			issue_number: Default::default(),
 			gate: Default::default(),
-		},
+		}
+		.into(),
 		timestamp: Utc::now().naive_utc(),
 		deduplication_id: "dedup1".to_string(),
 		origin: EventOrigin::Starknet,
@@ -46,12 +49,13 @@ fn creation_event(contribution_id: ContributionId) -> StorableEvent<Contribution
 fn assigned_event(
 	contribution_id: ContributionId,
 	contributor_account_address: ContributorAccountAddress,
-) -> StorableEvent<Contribution> {
-	StorableEvent {
+) -> Event {
+	Event {
 		event: ContributionEvent::Assigned {
 			id: contribution_id,
 			contributor_account_address,
-		},
+		}
+		.into(),
 		timestamp: Utc::now().naive_utc(),
 		deduplication_id: "dedup2".to_string(),
 		origin: EventOrigin::Starknet,
@@ -65,30 +69,25 @@ fn assigned_event(
 	ignore = "infrastructure test"
 )]
 fn test_append_and_list(
-	event_store: Box<dyn EventStore<Contribution>>,
+	database: &Client,
 	contribution_id: ContributionId,
-	creation_event: StorableEvent<Contribution>,
-	assigned_event: StorableEvent<Contribution>,
+	creation_event: Event,
+	assigned_event: Event,
 ) {
 	assert!(
-		event_store
-			.append(
-				&contribution_id,
-				vec![creation_event.clone(), assigned_event.clone()]
-			)
-			.is_ok()
+		EventStore::<Contribution>::append(
+			database,
+			&contribution_id,
+			vec![creation_event.clone(), assigned_event.clone()]
+		)
+		.is_ok()
 	);
 
-	let contribution_events = event_store.list_by_id(&contribution_id).unwrap();
+	let contribution_events =
+		DomainEventStore::<Contribution>::list_by_id(database, &contribution_id).unwrap();
 	assert_eq!(contribution_events.len(), 2);
-	assert_eq!(
-		*contribution_events.first().unwrap(),
-		Event::Contribution(creation_event.event)
-	);
-	assert_eq!(
-		*contribution_events.last().unwrap(),
-		Event::Contribution(assigned_event.event)
-	);
+	assert_eq!(*contribution_events.first().unwrap(), creation_event.event);
+	assert_eq!(*contribution_events.last().unwrap(), assigned_event.event);
 }
 
 #[rstest]
@@ -97,20 +96,21 @@ fn test_append_and_list(
 	ignore = "infrastructure test"
 )]
 fn test_cannot_append_duplicate_event_in_same_batch(
-	event_store: Box<dyn EventStore<Contribution>>,
+	database: &Client,
 	contribution_id: ContributionId,
-	creation_event: StorableEvent<Contribution>,
+	creation_event: Event,
 ) {
 	assert!(
-		event_store
-			.append(
-				&contribution_id,
-				vec![creation_event.clone(), creation_event]
-			)
-			.is_err()
+		EventStore::<Contribution>::append(
+			database,
+			&contribution_id,
+			vec![creation_event.clone(), creation_event]
+		)
+		.is_err()
 	);
 
-	let contribution_events = event_store.list_by_id(&contribution_id).unwrap();
+	let contribution_events =
+		DomainEventStore::<Contribution>::list_by_id(database, &contribution_id).unwrap();
 	assert_eq!(contribution_events.len(), 0);
 }
 
@@ -120,12 +120,22 @@ fn test_cannot_append_duplicate_event_in_same_batch(
 	ignore = "infrastructure test"
 )]
 fn test_cannot_append_duplicate_event_in_different_batches(
-	event_store: Box<dyn EventStore<Contribution>>,
+	database: &Client,
 	contribution_id: ContributionId,
-	creation_event: StorableEvent<Contribution>,
+	creation_event: Event,
 ) {
-	assert!(event_store.append(&contribution_id, vec![creation_event.clone()]).is_ok());
-	assert!(event_store.append(&contribution_id, vec![creation_event]).is_err());
+	assert!(
+		EventStore::<Contribution>::append(
+			database,
+			&contribution_id,
+			vec![creation_event.clone()]
+		)
+		.is_ok()
+	);
+	assert!(
+		EventStore::<Contribution>::append(database, &contribution_id, vec![creation_event])
+			.is_err()
+	);
 }
 
 #[rstest]
@@ -134,13 +144,16 @@ fn test_cannot_append_duplicate_event_in_different_batches(
 	ignore = "infrastructure test"
 )]
 fn test_metadata_are_stored(
+	database: &Client,
 	contribution_id: ContributionId,
-	creation_event: StorableEvent<Contribution>,
+	creation_event: Event,
 ) {
-	let client = Client::new(init_pool());
-	assert!(client.append(&contribution_id, vec![creation_event]).is_ok());
+	assert!(
+		EventStore::<Contribution>::append(database, &contribution_id, vec![creation_event])
+			.is_ok()
+	);
 
-	let connection = client.connection().unwrap();
+	let connection = database.connection().unwrap();
 	let list: Vec<(String, Option<Value>)> = events::dsl::events
 		.select((events::dsl::origin, events::dsl::metadata))
 		.load(&*connection)
