+++ marketplace-infrastructure/src/database/event_store/mod.rs
@@ -1,6 +1,3 @@
 #[allow(clippy::module_inception)]
 mod event_store;
 pub use event_store::*;
-
-#[cfg(test)]
-mod test;
